Add basis structure and layout with SPA
Implemented basic HTML and CSS structure with a home page and two side pages. The website constists of a single page application (SPA) that replaces the main content section of the home page.
The page contains a header bar with place for a company logo and a navigation menu. The main content section of the page is centered in the middle and its contents will be replaced. The footer is placed properly and sticks at the end of the page.