Implement API GET request to retieve timezone
The timezone for a city is retrieved by attaching a paramater 'city' to the GET request.

The server first looks up the city by the String given by the client. If the city is found, the next step is to look up the timezone for this city. Finally, the server sends the timezone utc string to the client.

Currently the server response is shown on the index.html page.