Add functionality of multiple homescreen clocks
The home page can display more than one clock at once.
The information about the clock is stored in arrays containing the canvas id and measurments(canvas, radius, ctx).
When the page is loaded the clock is drawn and calls a method every second to redraw the canvas.

The layout of the homepage is improved with a title, timesettings and clock face (image).