Clocks automatically generated with correct UTC
The inner HTML that displays the homescreen clocks is automatically generated with each city entry, with help of the server response.
The get request to the server looks up the name of the city and then continues to look up the timzeone (UTC) and gives this as a response to the client.
Based on only the city name(String), the clocks will be correctly displayed with the name, current time and UTC timzeone.

The process of generating and displaying the clock canvas are all done now in a separate javasript file, home.js.