Add button to change HTML contents
Addition of a button to change the inner HTML contents of the home page.
This is the basis of changing the homescreen clocks between analog and digital.
The javascript checks for a variable value that is either set to 'Analog' or 'Digital' and changes its state to the other value when the button is clicked. The inital value shall be retrieved from the users settings in the future. For now it remains a (String) value pre set in the code.