Layout update with addition to about page
Major layour update containing implementation of the color pallete to all pages. This is achieved by setting variables in CSS :root.
Other changes include company logo and background image.
Extra visual additions to cursor and art piece.

The about page contain a short discription of the project with the link to the github repo.