Improve settings page functionallity
Functionallity for display time type switch for changing the settings with server call. The switch displays the currect pre set time type and changes the settings directly.
When a city is added to the settings (server call), the table is updated accordingly.

The homescreen page is set with retrieving the settings from the server every time the page is reloaded. This is on first load, as well as when switched between settings/about page and home page. The settings are pre set time type and cities with timezones. When settings are changed they will always be updated on the home screen.